+++ src/TALib.NETCore.HighPerf/Candlestick/AdvanceBlock.cs
@@ -5,10 +5,10 @@ namespace TALib.NETCore.HighPerf
     public static partial class Lib
     {
         public static RetCode AdvanceBlock(
-            ref Span<double> inOpen,
-            ref Span<double> inHigh,
-            ref Span<double> inLow,
-            ref Span<double> inClose,
+            ref Span<decimal> inOpen,
+            ref Span<decimal> inHigh,
+            ref Span<decimal> inLow,
+            ref Span<decimal> inClose,
             int startIdx,
             int endIdx,
             int[] outInteger,
@@ -46,7 +46,7 @@ namespace TALib.NETCore.HighPerf
             int shadowLongTrailingIdx = startIdx - CandleAvgPeriod(CandleSettingType.ShadowLong);
             int nearTrailingIdx = startIdx - CandleAvgPeriod(CandleSettingType.Near);
             int farTrailingIdx = startIdx - CandleAvgPeriod(CandleSettingType.Far);
-            double bodyLongPeriodTotal = default;
+            decimal bodyLongPeriodTotal = default;
             int bodyLongTrailingIdx = startIdx - CandleAvgPeriod(CandleSettingType.BodyLong);
             int i = shadowShortTrailingIdx;
             while (i < startIdx)
