+++ src/TALib.NETCore.HighPerf/Candlestick/Hammer.cs
@@ -5,10 +5,10 @@ namespace TALib.NETCore.HighPerf
     public static partial class Lib
     {
         public static RetCode Hammer(
-            ref Span<double> inOpen,
-            ref Span<double> inHigh,
-            ref Span<double> inLow,
-            ref Span<double> inClose,
+            ref Span<decimal> inOpen,
+            ref Span<decimal> inHigh,
+            ref Span<decimal> inLow,
+            ref Span<decimal> inClose,
             int startIdx,
             int endIdx,
             int[] outInteger,
@@ -38,13 +38,13 @@ namespace TALib.NETCore.HighPerf
                 return RetCode.Success;
             }
 
-            double bodyPeriodTotal = default;
+            decimal bodyPeriodTotal = default;
             int bodyTrailingIdx = startIdx - CandleAvgPeriod(CandleSettingType.BodyShort);
-            double shadowLongPeriodTotal = default;
+            decimal shadowLongPeriodTotal = default;
             int shadowLongTrailingIdx = startIdx - CandleAvgPeriod(CandleSettingType.ShadowLong);
-            double shadowVeryShortPeriodTotal = default;
+            decimal shadowVeryShortPeriodTotal = default;
             int shadowVeryShortTrailingIdx = startIdx - CandleAvgPeriod(CandleSettingType.ShadowVeryShort);
-            double nearPeriodTotal = default;
+            decimal nearPeriodTotal = default;
             int nearTrailingIdx = startIdx - 1 - CandleAvgPeriod(CandleSettingType.Near);
             int i = bodyTrailingIdx;
             while (i < startIdx)
