+++ src/TALib.NETCore.HighPerf/Candlestick/StalledPattern.cs
@@ -5,10 +5,10 @@ namespace TALib.NETCore.HighPerf
     public static partial class Lib
     {
         public static RetCode StalledPattern(
-            ref Span<double> inOpen,
-            ref Span<double> inHigh,
-            ref Span<double> inLow,
-            ref Span<double> inClose,
+            ref Span<decimal> inOpen,
+            ref Span<decimal> inHigh,
+            ref Span<decimal> inLow,
+            ref Span<decimal> inClose,
             int startIdx,
             int endIdx,
             int[] outInteger,
@@ -41,9 +41,9 @@ namespace TALib.NETCore.HighPerf
             var bodyLongPeriodTotal = BufferHelpers.New(3);
             var nearPeriodTotal = BufferHelpers.New(3);
             int bodyLongTrailingIdx = startIdx - CandleAvgPeriod(CandleSettingType.BodyLong);
-            double bodyShortPeriodTotal = default;
+            decimal bodyShortPeriodTotal = default;
             int bodyShortTrailingIdx = startIdx - CandleAvgPeriod(CandleSettingType.BodyShort);
-            double shadowVeryShortPeriodTotal = default;
+            decimal shadowVeryShortPeriodTotal = default;
             int shadowVeryShortTrailingIdx = startIdx - CandleAvgPeriod(CandleSettingType.ShadowVeryShort);
             int nearTrailingIdx = startIdx - CandleAvgPeriod(CandleSettingType.Near);
             int i = bodyLongTrailingIdx;
