+++ src/TALib.NETCore.HighPerf/Candlestick/Tristar.cs
@@ -5,10 +5,10 @@ namespace TALib.NETCore.HighPerf
     public static partial class Lib
     {
         public static RetCode Tristar(
-            ref Span<double> inOpen,
-            ref Span<double> inHigh,
-            ref Span<double> inLow,
-            ref Span<double> inClose,
+            ref Span<decimal> inOpen,
+            ref Span<decimal> inHigh,
+            ref Span<decimal> inLow,
+            ref Span<decimal> inClose,
             int startIdx,
             int endIdx,
             int[] outInteger,
@@ -38,7 +38,7 @@ namespace TALib.NETCore.HighPerf
                 return RetCode.Success;
             }
 
-            double bodyPeriodTotal = default;
+            decimal bodyPeriodTotal = default;
             int bodyTrailingIdx = startIdx - 2 - CandleAvgPeriod(CandleSettingType.BodyDoji);
             int i = bodyTrailingIdx;
             while (i < startIdx - 2)
